+++ History.md
@@ -1,6 +1,7 @@
 ## v0.9.2.7 “ProfileScope”
 
 * Show user comments on user profile page. 
+* Move votes to their own `user.votes` object.
 
 ## v0.9.2.6 “InviteScope”